+++ src/store.ts
@@ -1,32 +1,15 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-import GithubData from '@/services/github-data';
-import { VuexModule, mutation, action, getter, Module } from 'vuex-class-component';
-import { GithubUser } from './models/github-user';
+import { user, UserStore } from './modules/User.vuex';
 
 Vue.use(Vuex);
 
-@Module( { namespacedPath: 'user/'})
-export class UserStore extends VuexModule {
-
-  // State/Getters
-  @getter public user: Partial<GithubUser> = {};
-
-  @action()
-  public async loadUser() {
-    this.setUser(await GithubData.getUser());
-  }
-
-  @mutation
-  private setUser(u: Partial<GithubUser>) {
-    this.user = u;
-  }
-}
-
-export const user = UserStore.ExtractVuexModule( UserStore );
-
-export default new Vuex.Store({
+export const store = new Vuex.Store({
   modules: {
     user,
   },
 });
+
+export const vxm = {
+  user: UserStore.CreateProxy( store, UserStore ) as UserStore,
+};
